DEMO_T02_V2/requirements.txt
Flask
Flask-RESTful
PyJWT
Flask-HTTPAuth